Added a new bib2tpl token
We now have a notion of `globalkey', which can be used to provide
a global key on each invocation of publist. For me this allows for
one template to be used multiple time within the same HTML page scope,
but with each invocation creating a unique ID. This makes Javascript
related tasks much easier.